benchmark tables

+++ scripts/benchmark_python.sh
@@ -1,31 +1,60 @@
 #!/bin/bash
 
-rm -f results/python-mip/cbc_cpython.csv results/python-mip/gurobi_cpython.csv results/python-mip/cbc_pypy.csv results/python-mip/gurobi_pypy.csv results/gurobipy.csv results/pulp_cpython.csv results/pulp_pypy.csv
+mkdir -p results/python-mip
 
-echo "Benchmarking n-Queens CBC Python-MIP CPYTHON"
-python3 python-mip/benchmarks/queens.py cbc
-mv queens-mip-cbc.csv results/python-mip/cbc_cpython.csv
+if [ ! -f results/python-mip/cbc_cpython.csv ]; then
+    echo "Benchmarking n-Queens CBC Python-MIP CPYTHON"
+    python3 python-mip/benchmarks/queens.py cbc && \
+        sed -i '1s/^/N,num_variables,num_constraints,num_entries,model_time_s\n/' queens-mip-cbc.csv && \
+        mv queens-mip-cbc.csv results/python-mip/cbc_cpython.csv
+    rm queens-mip-cbc.csv
+fi
 
-echo "Benchmarking n-Queens Gurobi Python-MIP CPYTHON"
-python3 python-mip/benchmarks/queens.py gurobi
-mv queens-mip-gurobi.csv results/python-mip/gurobi_cpython.csv
+if [ ! -f results/python-mip/gurobi_cpython.csv ]; then
+    echo "Benchmarking n-Queens Gurobi Python-MIP CPYTHON"
+    python3 python-mip/benchmarks/queens.py gurobi && \
+        sed -i '1s/^/N,num_variables,num_constraints,num_entries,model_time_s\n/' queens-mip-gurobi.csv && \
+        mv queens-mip-gurobi.csv results/python-mip/gurobi_cpython.csv
+    rm queens-mip-gurobi.csv
+fi
 
-echo "Benchmarking n-Queens CBC Python-MIP Pypy"
-pypy3 python-mip/benchmarks/queens.py cbc
-mv queens-mip-cbc.csv results/python-mip/cbc_pypy.csv
+if [ ! -f results/python-mip/cbc_pypy.csv ]; then
+    echo "Benchmarking n-Queens CBC Python-MIP Pypy"
+    pypy3 python-mip/benchmarks/queens.py cbc && \
+        sed -i '1s/^/N,num_variables,num_constraints,num_entries,model_time_s\n/' queens-mip-cbc.csv && \
+        mv queens-mip-cbc.csv results/python-mip/cbc_pypy.csv
+    rm queens-mip-cbc.csv
+fi
 
-echo "Benchmarking n-Queens Gurobi Python-MIP Pypy"
-pypy3 python-mip/benchmarks/queens.py gurobi
-mv queens-mip-gurobi.csv results/python-mip/gurobi_pypy.csv
+if [ ! -f results/python-mip/gurobi_pypy.csv ]; then
+    echo "Benchmarking n-Queens Gurobi Python-MIP Pypy"
+    pypy3 python-mip/benchmarks/queens.py gurobi && \
+        sed -i '1s/^/N,num_variables,num_constraints,num_entries,model_time_s\n/' queens-mip-gurobi.csv && \
+        mv queens-mip-gurobi.csv results/python-mip/gurobi_pypy.csv
+    rm queens-mip-gurobi.csv
+fi
 
-echo "Benchmarking n-Queens Gurobi"
-python3 python-mip/benchmarks/queens-gurobi.py
-mv queens-gurobi-cpython.csv results/gurobipy.csv
+if [ ! -f results/gurobipy.csv ]; then
+    echo "Benchmarking n-Queens Gurobi"
+    python3 python-mip/benchmarks/queens-gurobi.py && \
+        mv queens-gurobi-cpython.csv results/gurobipy.csv
+    rm queens-gurobi-cpython.csv
+fi
 
-echo "Benchmarking n-Queens Pulp CPYTHON"
-python3 python-mip/benchmarks/queens-pulp.py
-mv queens-pulp.csv results/pulp_cpython.csv
+mkdir -p results/pulp
 
-echo "Benchmarking n-Queens Pulp Pypy"
-pypy3 python-mip/benchmarks/queens-pulp.py
-mv queens-pulp.csv results/pulp_pypy.csv
+if [ ! -f results/pulp/cpython.csv ]; then
+    echo "Benchmarking n-Queens Pulp CPYTHON"
+    python3 python-mip/benchmarks/queens-pulp.py && \
+        sed -i '1s/^/N,name,model_time_s\n/' queens-pulp.csv && \
+        mv queens-pulp.csv results/pulp/cpython.csv
+    rm queens-pulp.csv
+fi
+
+if [ ! -f results/pulp/pypy.csv ]; then
+    echo "Benchmarking n-Queens Pulp Pypy"
+    pypy3 python-mip/benchmarks/queens-pulp.py && \
+        sed -i '1s/^/N,name,model_time_s\n/' queens-pulp.csv && \
+        mv queens-pulp.csv results/pulp/pypy.csv
+    rm queens-pulp.csv
+fi
